included fmt headers, removed dword usage on non windows specific funcs

+++ assets/arcane/core/sdk/ifaces/player_info.hh
@@ -22,7 +22,7 @@ namespace arcane::sdk::ifaces {
 
 		PAD( pad, 16 );
 
-		DWORD friends_id;
+		unsigned long friends_id;
 
 		char friends_name[ 128 ];
